Refactored promise.js, removed validator functions Relates #41
